adding a new test for watch with namespace
and factoring the test logic into a common function
